revert `test_store_backends.py` move
